refactor(styles): theme export improved
