+++ libs/components/CharacterButtons/index.tsx
@@ -3,7 +3,7 @@ import { characterButtons } from '../../types';
 import React, { FC, Fragment } from 'react';
 import Link from '../Link';
 import { DOMAINS } from '../../constants';
-import { theme } from '../../styles/theme';
+import { theme } from '../../styles';
 
 const styleCss = {
   large: {
