+++ pages/_document.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Document, { Html, Head, Main, NextScript } from 'next/document';
 import { ServerStyleSheets } from '@material-ui/core/styles';
-import { theme } from '../libs/styles/theme';
+import { theme } from '../libs/styles';
 
 export default class MyDocument extends Document {
   render() {
